requirements.txt
pandas
sqlalchemy
psycopg2
openai
streamlit
python-dotenv
matplotlib
seaborn
scikit-learn